bot/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.1"
tgcrypto = "^1.2.5"
pyrogram = { git = "https://github.com/KurimuzonAkuma/pyrogram.git", tag = "v2.1.31" }
pyrostates = { git = "https://github.com/Scorpi-ON/pyrostates.git" }
uvloop = { version = "^0.21.0", markers = "sys_platform == 'linux'" }
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
pony = "^0.7.19"
httpx = "^0.28.0"
pydantic = {extras = ["email"], version = "^2.10.3"}
python-multipart = "^0.0.19"

[tool.poetry.group.dev.dependencies]
ruff = "0.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
